First pass client refresh

+++ src/provider/index.ts
@@ -3,21 +3,24 @@ import {ChannelProvider} from 'openfin/_v2/api/interappbus/channel/provider';
 
 import {CHANNEL_NAME} from '../shared/config';
 import {Notification} from '../shared/models/Notification';
-import {NotificationEvent} from '../shared/models/NotificationEvent';
-import {NotificationTypes, TypeResolver} from '../shared/models/NotificationTypes';
+import {NotificationEvent, NotificationClickedEvent, NotificationButtonClickedEvent, NotificationClosedEvent} from '../client/models/NotificationEvent';
+import {NotificationType, resolveType} from '../shared/models/NotificationTypes';
 
 import {ISenderInfo} from './models/ISenderInfo';
 import {HistoryRepository} from './persistence/dataLayer/repositories/HistoryRepository';
 import {Repositories} from './persistence/dataLayer/repositories/RepositoryEnum';
 import {RepositoryFactory} from './persistence/dataLayer/repositories/RepositoryFactory';
 import {SettingsRepository} from './persistence/dataLayer/repositories/SettingsRepository';
+import { APIHandler } from './APIHandler';
+import { APITopic, API, CreatePayload, ClearPayload } from '../client/internal';
+import { OptionButton, OptionInput } from '../client/models/NotificationOptions';
+import { Identity } from 'openfin/_v2/main';
 
 const repositoryFactory = RepositoryFactory.Instance;
 const historyRepository = repositoryFactory.getRepository(Repositories.history) as HistoryRepository;
 const settingsRepository = repositoryFactory.getRepository(Repositories.settings) as SettingsRepository;
 
 let providerChannel: ChannelProvider;
-let serviceUUID: string;
 const centerIdentity = {
     name: 'Notification-Center',
     uuid: '',
@@ -61,9 +64,19 @@ fin.desktop.main(() => {
  */
 async function registerService() {
     // Register the service
+    const apiHandler: APIHandler<APITopic> = new APIHandler();
+
+    await apiHandler.registerListeners<API>({
+        [APITopic.CREATE_NOTIFICATION]: createNotification,
+        [APITopic.CLEAR_NOTIFICATION]: clearNotification,
+        [APITopic.GET_APP_NOTIFICATIONS]: fetchAppNotifications,
+        [APITopic.CLEAR_APP_NOTIFICATIONS]: clearAppNotifications,
+        [APITopic.TOGGLE_NOTIFICATION_CENTER]: toggleNotificationCenter
+    });
+
     const serviceId = fin.desktop.Application.getCurrent().uuid;
-    providerChannel = await fin.InterApplicationBus.Channel.create(CHANNEL_NAME);
-    centerIdentity.uuid = serviceUUID = serviceId;
+    providerChannel = apiHandler.channel; // Temporary while I clean out all of the other references to this
+    centerIdentity.uuid = serviceId;
 
     // handle client connections
     providerChannel.onConnection((app, payload) => {
@@ -74,21 +87,16 @@ async function registerService() {
         }
     });
 
-    // Functions called by the client
-    providerChannel.register('create-notification', createNotification);
-    providerChannel.register('toggle-notification-center', toggleNotificationCenter);
-
-    // Functions called either by the client or the Notification Center
-    providerChannel.register('clear-notification', clearNotification);
-    providerChannel.register('fetch-app-notifications', fetchAppNotifications);
-    providerChannel.register('clear-app-notifications', clearAppNotifications);
-
     // Functions called by the Notification Center
-    providerChannel.register('notification-clicked', notificationClicked);
-    providerChannel.register('notification-button-clicked', notificationButtonClicked);
-    providerChannel.register('notification-closed', notificationClosed);
-    providerChannel.register('fetch-all-notifications', fetchAllNotifications);
-    providerChannel.register('clear-all-notifications', clearAllNotifications);
+    apiHandler.channel.register('notification-clicked', notificationClicked);
+    apiHandler.channel.register('notification-button-clicked', notificationButtonClicked);
+    apiHandler.channel.register('notification-closed', notificationClosed);
+    apiHandler.channel.register('fetch-all-notifications', fetchAllNotifications);
+    apiHandler.channel.register('clear-all-notifications', clearAllNotifications);
+}
+
+async function dispatchClientEvent(target: Identity, payload: NotificationEvent): Promise<void> {
+    return providerChannel.dispatch(target, 'event', payload);
 }
 
 // Send notification created to the UI
@@ -112,30 +120,6 @@ const dispatchAppNotificationsCleared = async (payload: {uuid: string}) => {
     console.log('success', success);
 };
 
-// Send notification clicked to the Client
-const dispatchNotificationClicked = async (payload: NotificationEvent) => {
-    const providerChannelPlugin = await providerChannel;
-    const clientIdentity = {name: payload.name, uuid: payload.uuid, channelId: '', channelName: ''};
-    const success = await providerChannelPlugin.dispatch(clientIdentity, 'notification-clicked', payload);
-    console.log('success', success);
-};
-
-// Send notification clicked to the Client
-const dispatchNotificationButtonClicked = async (payload: Notification&ISenderInfo&{buttonIndex: number}) => {
-    const providerChannelPlugin = await providerChannel;
-    const clientIdentity = {name: payload.name, uuid: payload.uuid, channelId: '', channelName: ''};
-    const success = await providerChannelPlugin.dispatch(clientIdentity, 'notification-button-clicked', payload);
-    console.log('success', success);
-};
-
-// Send notification closed to the Client
-const dispatchNotificationClosed = async (payload: NotificationEvent) => {
-    const providerChannelPlugin = await providerChannel;
-    const clientIdentity = {name: payload.name, uuid: payload.uuid, channelId: '', channelName: ''};
-    const success = await providerChannelPlugin.dispatch(clientIdentity, 'notification-closed', payload);
-    console.log('success', success);
-};
-
 const dispatchAllNotificationsCleared = async (payload: Notification) => {
     const providerChannelPlugin = await providerChannel;
     const success = await providerChannelPlugin.dispatch(centerIdentity, 'all-notifications-cleared', payload);
@@ -148,12 +132,11 @@ const dispatchToggleNotificationCenter = async (payload: undefined) => {
     console.log('success', success);
 };
 
-
 /**
  * Encodes the Id which currently is the uuid:id
  * @param payload The notification object
  */
-function encodeID(payload: Notification&ISenderInfo|NotificationEvent|{id: string}&ISenderInfo): string {
+function encodeID(payload: Notification&ISenderInfo|{id: string}&ISenderInfo): string {
     return `${payload.uuid}:${payload.id}`;
 }
 
@@ -161,7 +144,7 @@ function encodeID(payload: Notification&ISenderInfo|NotificationEvent|{id: strin
  * @method decodeID This function retrieves the notification Id
  * @param payload The notification object
  */
-function decodeID(payload: Notification&ISenderInfo|NotificationEvent|{id: string}&ISenderInfo): string {
+function decodeID(payload: Notification&ISenderInfo|{id: string}&ISenderInfo): string {
     return payload.id.slice(payload.uuid.length + 1);
 }
 
@@ -170,18 +153,29 @@ function decodeID(payload: Notification&ISenderInfo|NotificationEvent|{id: strin
  * @param {object} payload The contents to be dispatched to the UI
  * @param {object} sender Window info for the sending client. This can be found in the relevant app.json within the demo folder.
  */
-async function createNotification(payload: Notification, sender: ProviderIdentity) {
+async function createNotification(payload: CreatePayload, sender: ProviderIdentity): Promise<Notification> {
     // For testing/display purposes
     console.log('createNotification hit');
 
     console.log('payload', payload);
     console.log('sender', sender);
 
-    testDisplay('createNotification', payload, sender);
+    const notification: Notification = {
+        id: payload.id,
+        body: payload.body,
+        title: payload.title,
+        subtitle: payload.subtitle || '',
+        icon: payload.icon || '',
+        context: payload.context,
+        date: payload.date || new Date(),
+        buttons: payload.buttons || [] as OptionButton[],
+        inputs: payload.inputs || [] as OptionInput[],
+        type: resolveType(payload)
+    };
 
-    const noteType: NotificationTypes = TypeResolver(payload);
-    const fullPayload: Notification&ISenderInfo = Object.assign({}, payload, sender as ISenderInfo, {type: noteType});
-    const encodedID: string = encodeID(fullPayload);
+    testDisplay('createNotification', payload, sender);
+    const fullPayload: Notification&ISenderInfo = Object.assign({}, notification, sender as ISenderInfo);
+    const encodedID: string = `${sender.uuid}:${payload.id}`;
 
     const fullPayloadEncoded: Notification&ISenderInfo = Object.assign({}, fullPayload, {id: encodedID});
 
@@ -190,19 +184,20 @@ async function createNotification(payload: Notification, sender: ProviderIdentit
 
     // Create the notification/toast in the UI
     if (result.success) {
-        dispatchNotificationCreated(fullPayload);
+        await dispatchNotificationCreated(fullPayloadEncoded);
+        // Return the full notifiation to the client
+        return notification;
+    } else {
+        throw new Error(result.errorMsg);
     }
 
-    // Return a notification event with the notification context/id to the client
-    // that created the notification Or return the payload?
-    return result;
 }
 
 /**
  * @method toggleNotificationCenter Dispatch the notification center toggle action to the UI
  * @param {object} sender Window info for the sending client. This can be found in the relevant app.json within the demo folder.
  */
-async function toggleNotificationCenter(payload: undefined, sender: ProviderIdentity) {
+async function toggleNotificationCenter(payload: undefined, sender: ProviderIdentity): Promise<void> {
     // For testing/display purposes
     console.log('toggleNotificationCenter hit');
 
@@ -210,9 +205,7 @@ async function toggleNotificationCenter(payload: undefined, sender: ProviderIden
     console.log('sender', sender);
 
     // Toggle the notification center
-    dispatchToggleNotificationCenter(undefined);
-
-    return '';
+    return dispatchToggleNotificationCenter(undefined);
 }
 
 /**
@@ -221,7 +214,7 @@ async function toggleNotificationCenter(payload: undefined, sender: ProviderIden
  * @param {object} sender Window info for the sending client. This can be found in the relevant app.json within the demo folder.
  * @returns {ReturnResult} Whether or not the removal of the notifications was successful.
  */
-async function clearNotification(payload: Notification, sender: ProviderIdentity) {
+async function clearNotification(payload: ClearPayload, sender: ProviderIdentity): Promise<void> {
     // For testing/display purposes
     console.log('clearNotification hit');
 
@@ -232,21 +225,25 @@ async function clearNotification(payload: Notification, sender: ProviderIdentity
 
     // Grab notificationID from payload. If not present, return error?
     if (!payload.id) {
-        return 'ERROR: Must supply a notification ID to clear!';
+        throw new Error('Must supply a notification ID to clear!');
     }
 
-    const fullPayload: Notification&ISenderInfo = Object.assign({}, payload, sender as ISenderInfo);
-    const encodedID: string = encodeID(fullPayload);
+    const encodedID: string = `${sender.uuid}:${payload.id}`;
+    const retrievalResult = await historyRepository.getById(encodedID);
+    if (!retrievalResult.success) {
+        throw new Error(retrievalResult.errorMsg);
+    }
+    const fullPayload: Notification&ISenderInfo = Object.assign({}, retrievalResult.value, sender as ISenderInfo);
 
     // Delete notification from indexeddb
     const result = await historyRepository.remove(encodedID);
 
     // Delete the notification/toast in UI
     if (result.success) {
-        dispatchNotificationCleared(fullPayload);
+        await dispatchNotificationCleared(fullPayload);
+    } else {
+        throw new Error(result.errorMsg);
     }
-
-    return result;
 }
 
 /**
@@ -254,7 +251,7 @@ async function clearNotification(payload: Notification, sender: ProviderIdentity
  * @param {object} payload Should contain the id of the notification clicked. Also the uuid and name of the original Client window.
  * @param {object} sender UI Window info. Not important.
  */
-function notificationClicked(payload: NotificationEvent&ISenderInfo, sender: ProviderIdentity) {
+async function notificationClicked(payload: NotificationEvent&ISenderInfo, sender: ProviderIdentity): Promise<void> {
     // For testing/display purposes
     console.log('notificationClicked hit');
 
@@ -263,15 +260,15 @@ function notificationClicked(payload: NotificationEvent&ISenderInfo, sender: Pro
 
     testDisplay('notificationClicked', payload, sender);
 
+    const target: Identity = {uuid: payload.uuid, name: payload.name};
+    const event: NotificationClickedEvent = {type: 'notification-clicked', id: payload.id};
+    
     // Send notification clicked event to uuid with the context.
-    dispatchNotificationClicked(payload);
-
-    // TODO: What should we return?
-    return 'notificationClicked returned';
+    dispatchClientEvent(target, event);
 }
 
 
-function notificationButtonClicked(payload: Notification&ISenderInfo&{buttonIndex: number}, sender: ProviderIdentity) {
+async function notificationButtonClicked(payload: Notification&ISenderInfo&{buttonIndex: number}, sender: ProviderIdentity): Promise<void> {
     // For testing/display purposes
     console.log('notificationButtonClicked hit');
 
@@ -280,11 +277,10 @@ function notificationButtonClicked(payload: Notification&ISenderInfo&{buttonInde
 
     testDisplay('notificationButtonClicked', payload, sender);
 
-    // Send notification clicked event to uuid with the context.
-    dispatchNotificationButtonClicked(payload);
+    const target: Identity = {uuid: payload.uuid, name: payload.name};
+    const event: NotificationButtonClickedEvent = {type: 'notification-button-clicked', id: payload.id, button: payload.buttonIndex};
 
-    // TODO: What should we return?
-    return 'notificationButtonClicked returned';
+    dispatchClientEvent(target, event);
 }
 
 /**
@@ -293,7 +289,7 @@ function notificationButtonClicked(payload: Notification&ISenderInfo&{buttonInde
  * @param {object} payload Should contain the id of the notification clicked. Also the uuid and name of the original Client window.
  * @param {object} sender UI Window info. Not important.
  */
-async function notificationClosed(payload: Notification&ISenderInfo&{buttonIndex: number}, sender: ProviderIdentity) {
+async function notificationClosed(payload: Notification&ISenderInfo&{buttonIndex: number}, sender: ProviderIdentity): Promise<void> {
     // For testing/display purposes
     console.log('notificationClosed hit');
 
@@ -309,12 +305,14 @@ async function notificationClosed(payload: Notification&ISenderInfo&{buttonIndex
 
     // Send notification closed event to uuid with the context.
     if (result.success) {
-        dispatchNotificationCleared(payload);
-        dispatchNotificationClosed(payload);
-    }
+        const target: Identity = {uuid: payload.uuid, name: payload.name};
+        const event: NotificationClosedEvent = {type: 'notification-closed', id: payload.id};
 
-    // TODO: What should we return?
-    return result;
+        dispatchNotificationCleared(payload);
+        dispatchClientEvent(target, event);
+    } else {
+        throw new Error(result.errorMsg);
+    }
 }
 
 /**
@@ -354,7 +352,7 @@ async function fetchAllNotifications(payload: undefined, sender: ProviderIdentit
  * @param {undefined} payload The payload can contain the uuid
  * @param {ISenderInfo} sender The sender info contains the uuid of the sender
  */
-async function fetchAppNotifications(payload: NotificationEvent, sender: ProviderIdentity) {
+async function fetchAppNotifications(payload: undefined , sender: ProviderIdentity): Promise<Notification[]> {
     // For testing/display purposes
     console.log('fetchAppNotifications hit');
 
@@ -363,16 +361,9 @@ async function fetchAppNotifications(payload: NotificationEvent, sender: Provide
 
     testDisplay('fetchAppNotifications', payload, sender);
 
-    // If no uuid supplied in the payload, set to the sender's uuid. This is so
-    // you can request another app's notification (such as for the Notification
-    // Center)
-    if (!payload.uuid) {
-        payload.uuid = sender.uuid;
-    }
-
     // Grab an app's notifications from indexeddb
     // Send those notifications to the requesting app.
-    const result = await historyRepository.getByUuid(payload.uuid);
+    const result = await historyRepository.getByUuid(sender.uuid);
 
     if (result.success) {
         const appNotifications = result.value as (Notification & ISenderInfo)[];
@@ -382,10 +373,10 @@ async function fetchAppNotifications(payload: NotificationEvent, sender: Provide
             notification.id = decodeID(notification);
         });
 
-        result.value = appNotifications;
+        return appNotifications;
+    } else {
+        throw new Error(result.errorMsg);
     }
-
-    return result;
 }
 
 /**
@@ -413,7 +404,7 @@ async function clearAllNotifications(payload: Notification, sender: ProviderIden
     return 'clearAllNotifications returned';
 }
 
-function clearAppNotifications(payload: {uuid: string}, sender: ProviderIdentity) {
+async function clearAppNotifications(payload: undefined, sender: ProviderIdentity): Promise<void> {
     console.log('clearAppNotifications hit');
 
     console.log('payload', payload);
@@ -422,15 +413,16 @@ function clearAppNotifications(payload: {uuid: string}, sender: ProviderIdentity
     testDisplay('clearAppNotifications', payload, sender);
 
     // Delete app notifications from indexeddb
-    historyRepository.removeByUuid(payload.uuid);
+    const result = await historyRepository.removeByUuid(sender.uuid);
 
-    // Delete the notifications/toasts in UI (TODO: NEED TO REGISTER
-    // APP-NOTIFICATIONS-CLEARED IN UI)
-    dispatchAppNotificationsCleared(payload);
+    if (result.success) {
+        // Delete the notifications/toasts in UI (TODO: NEED TO REGISTER
+        // APP-NOTIFICATIONS-CLEARED IN UI)
+        await dispatchAppNotificationsCleared(sender);
+    } else {
+        throw new Error(result.errorMsg);
+    }
 
-
-    // TODO: What should we return?
-    return 'clearAppNotifications returned';
 }
 
 /**
@@ -439,7 +431,7 @@ function clearAppNotifications(payload: {uuid: string}, sender: ProviderIdentity
  * @param {Notification} payload The notification payload
  * @param {ISenderInfo} sender The sender info
  */
-function testDisplay(action: string, payload: Notification|NotificationEvent|{uuid: string}, sender: ProviderIdentity) {
+function testDisplay(action: string, payload: any, sender: ProviderIdentity) {
     document.body.innerHTML = '';
     document.body.innerHTML += `${action} hit <br></br>`;
     document.body.innerHTML += 'PAYLOAD <br></br>';
